requirements.txt
opencv-python
mediapipe
pyserial
SpeechRecognition
pyaudio
